+++ example/movies_app/test/movies_app_test.dart
@@ -1,7 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  testWidgets('Movie added to favorite', (WidgetTester tester) async {
-    // TODO implement
-  });
-}