+++ lib/src/base_bloc.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:bloc/bloc.dart';
 
-/// Bloc bas state
+/// Bloc base state
 /// All states must inherit from [BlocState]
 abstract class BlocState implements Equatable {
   @override
@@ -9,54 +9,55 @@ abstract class BlocState implements Equatable {
 }
 
 /// The initial state of any bloc that extends [BaseBloc]
-class StateInitial extends BlocState {}
+/// When the initial state is StateUninitialized, then the first event that
+/// will be dispatched is the [InitializeEvent] which will call
+/// [BaseBloc.initialize]
+class StateUninitialized extends BlocState {}
 
-/// Bloc loading state
-/// Each state that needs to allow loading functionality to be used with
-/// the [StateBuilder] must be implemented with [StateLoading]
-mixin StateLoading implements BlocState {}
+/// Loading state. Used when the bloc is performing some action...
+class StateLoading extends BlocState {
+  final double progress;
 
-/// Bloc error state
-/// Each state that needs tto allow error functionality to be used with the
-/// [StateBuilder[ must be implemented with [StateError]
-mixin StateError implements BlocState {
-  String get message;
+  StateLoading([this.progress]);
+
+  @override
+  List<Object> get props => [progress];
 }
 
 /// A state that indicates that the bloc is being initialized. Allows reporting
 /// [progress] of initializing
 /// This state is implemented with [StateLoading] so that it can be used as
 /// a loading state with the [StateBuilder]
-class StateInitializing with StateLoading {
-  final double progress;
-
-  StateInitializing([this.progress]);
-
-  @override
-  List<Object> get props => [progress];
+class StateInitializing extends StateLoading {
+  StateInitializing([double progress]) : super(progress);
 }
 
-/// Emitted when there is an issue with initializing the bloc. It contains
-/// the error [message], [exception] and [stacktrace]
-/// This state is implemented with [StateError] so that it can be used as an
-/// error state with the [StateBuilder]
-class StateInitializationError with StateError {
+/// Bloc error state
+/// Each state that needs tto allow error functionality to be used with the
+/// [StateBuilder[ must be implemented with [StateError]
+class StateError extends BlocState {
   final String message;
-  final Object exception;
-  final StackTrace stacktrace;
+  final dynamic exception;
+  final StackTrace stackTrace;
 
-  StateInitializationError(this.message, [this.exception, this.stacktrace]);
+  StateError(this.message, [this.exception, this.stackTrace]);
 
   @override
   List<Object> get props => [message, exception];
 }
 
+class StateInitializationError extends StateError {
+  StateInitializationError(String message,
+      [dynamic exception, StackTrace stackTrace])
+      : super(message, exception, stackTrace);
+}
+
 /// A state that indicates that the bloc is initialized
 class StateInitialized extends BlocState {}
 
-/// Base bloc event
+/// Bloc base event
 /// All events must inherit from [BlocEvent]
-class BlocEvent extends Equatable {
+abstract class BlocEvent implements Equatable {
   @override
   List<Object> get props => [];
 }
@@ -70,17 +71,25 @@ class InitializeEvent extends BlocEvent {}
 /// In order to do some initializations, you can override the [initialize] method
 /// If you need to yield some state upon initialization, then you can use the
 /// [onInitialized] method
-abstract class BaseBloc<E extends BlocEvent, S extends BlocState>
-    extends Bloc<BlocEvent, BlocState> {
+abstract class BaseBloc extends Bloc<BlocEvent, BlocState> {
   bool _initialized = false;
 
   BaseBloc() {
-    super.add(InitializeEvent());
+    if (initialState is StateUninitialized) {
+      super.add(InitializeEvent());
+    } else {
+      // if the initial state is not StateInitial, then the bloc will be
+      // initialized directly without calling the initialize() method.
+      _initialized = true;
+    }
   }
 
-  /// The initial state of the bloc is always StateInitial
+  /// The default initial state of the bloc is [StateUninitialized]
+  /// If you override [initialState] to some other state rather than
+  /// [StateUninitialized], then [InitializeEvent] won't be dispatched, and the
+  /// [initialize] method won't be called
   @override
-  BlocState get initialState => StateInitial();
+  BlocState get initialState => StateUninitialized();
 
   @override
   Stream<BlocState> mapEventToState(BlocEvent event) async* {
@@ -92,47 +101,30 @@ abstract class BaseBloc<E extends BlocEvent, S extends BlocState>
         yield StateInitialized();
         yield* onInitialized();
       } catch (e, stacktrace) {
-        yield StateInitializationError('Initialization error', e, stacktrace);
+        yield StateError('Initialization error', e, stacktrace);
       }
     } else if (_initialized) {
-      assert(event is E);
-      yield* eventToState(event as E);
+      yield* eventToState(event);
     } else {
       throw Exception('Bloc is not initialized yet...');
     }
   }
 
-  @override
-  @deprecated
-
-  /// This method is deprecated. Use the [event] method instead for adding
-  /// events to the bloc
-  void add(BlocEvent e) {
-    throw Exception(
-        "Using the add method is not allowed. Use the event() method instead");
-  }
-
-  /// Adds a new event to the stream
-  /// Must be used instead of the add() method of the [Bloc] class
-  void event(E event) {
-    super.add(event);
-  }
-
   /// Used for initializing the block.
   /// Can update progress by yielding StateInitializing.
   /// If you want to yield an error, then throw an exception.
-  /// If an exception is thrown, then a [StateInitializationError] is yielded
+  /// If an exception is thrown, then a [StateError] is yielded
   Stream<StateInitializing> initialize() async* {}
 
   /// Called when the bloc is initialized (when [StateInitialized] is yielded
   /// Can be used for adding some event or for yielding some state after
   /// initialization is done
-  Stream<S> onInitialized() async* {}
+  Stream<BlocState> onInitialized() async* {}
 
   /// Must be implemented when a class extends [BaseBloc].
   /// Takes the incoming [event] as the argument.
   /// [eventToState] is called whenever an [event] is added.
   /// [eventToState] must convert that [event] into a new [state]
   /// and return the new [state] in the form of a `Stream<State>`.
-  Stream<S> eventToState(E event);
+  Stream<BlocState> eventToState(BlocEvent event);
 }