+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.1.2] - 2020-01-29
+
+* Simplified BaseBloc and ComponentView stateBuilder()
+
+## [0.1.1+1] - 2020-01-12
+
+* Updated diagrams links in the README
+
 ## [0.1.1] - 2020-01-12
 
 * Small fixes to the docs